// gen/ts/open_splunk/app.ts
export enum AppState {
  APP_STATE_UNSPECIFIED = 0,
  APP_STATE_ACTIVE = 1,
  APP_STATE_ARCHIVED = 2,
  UNRECOGNIZED = -1,
}